Moved ThemeInfoDialog to :core
