+++ components/src/main/res/navigation/nav_components.xml
@@ -285,7 +285,7 @@
         app:destination="@id/ThemeInfoDialog" />
     <dialog
         android:id="@+id/ThemeInfoDialog"
-        android:name="com.numero.material_gallery.themeinfo.ThemeInfoBottomSheetDialog"
+        android:name="com.numero.material_gallery.core.ThemeInfoBottomSheetDialog"
         tools:layout="@layout/bottom_sheet_fragment_theme_info" />
 
 </navigation>
