+++ core/src/main/res/color/stroke_color.xml
