+++ core/src/main/res/drawable/bottom_sheet_controll_bar.xml
