+++ core/src/main/res/layout/view_color_info_item.xml
