+++ core/src/main/res/layout/view_typography_info_item.xml
