+++ core/src/main/res/values/attrs.xml
