+++ resource/src/main/res/drawable/ic_android.xml
