+++ resource/src/main/res/drawable/ic_light.xml
