+++ resource/src/main/res/drawable/ic_night.xml
